+++ CHANGES.rst
@@ -7,6 +7,7 @@ development
 - Improve Celery bootstrapping
 - Add “--poll” option to wait for results when running in “--background” mode
 - Improve command line argument handling
+- Rename command line option "--poll" to "--wait"
 
 2017-10-21 0.3.0
 ================